Update send_encrypted_email.md
Thanks for your pull request. I think this is important enough to merit a note at the beginning of the topic. I've made a few small changes to align the text with AWS Style guidelines. Let me know if you this change looks good to you, and I'll merge the PR.